requirements.txt
fastapi==0.115.4
uvicorn==0.31.1
python-multipart==0.0.12
python-dotenv==1.0.1
openai==1.54.4
langchain
langchain-openai
langchain-tavily
langgraph
pydantic==2.9.2
httpx==0.27.2 
langchain-community